+++ Assets/VRM10/Editor/Components/Expression/Vrm10InstanceEditorExpression.cs
@@ -48,16 +48,16 @@ namespace UniVRM10
                 {
                     m_expressionKeyWeights[slider.Key] = slider.Value;
                 }
-                m_target.Vrm.Expression.SetWeights(m_expressionKeyWeights);
+                m_target.Runtime.Expression.SetWeights(m_expressionKeyWeights);
             }
 
             EditorGUILayout.Space();
             EditorGUILayout.LabelField("Override rates", EditorStyles.boldLabel);
             EditorGUI.BeginDisabledGroup(true);
             {
-                EditorGUILayout.Slider("Blink override rate", m_target.Vrm.Expression.BlinkOverrideRate, 0f, 1f);
-                EditorGUILayout.Slider("LookAt override rate", m_target.Vrm.Expression.LookAtOverrideRate, 0f, 1f);
-                EditorGUILayout.Slider("Mouth override rate", m_target.Vrm.Expression.MouthOverrideRate, 0f, 1f);
+                EditorGUILayout.Slider("Blink override rate", m_target.Runtime.Expression.BlinkOverrideRate, 0f, 1f);
+                EditorGUILayout.Slider("LookAt override rate", m_target.Runtime.Expression.LookAtOverrideRate, 0f, 1f);
+                EditorGUILayout.Slider("Mouth override rate", m_target.Runtime.Expression.MouthOverrideRate, 0f, 1f);
             }
             EditorGUI.EndDisabledGroup();
         }
