+++ Assets/VRM10/Runtime/Components/Vrm10Runtime/Vrm10RuntimeExpression.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c644583b448f6e743a6d2a55d1d617bf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
